+++ app/src/main/java/com/example/KC/ui/bottom/profile/ProfileViewModel.java
@@ -1,0 +1,7 @@
+package com.example.KC.ui.bottom.profile;
+
+import androidx.lifecycle.ViewModel;
+
+public class ProfileViewModel extends ViewModel {
+    // TODO: Implement the ViewModel
+}
