+++ app/src/main/java/com/example/KC/ui/bottom/search/SearchFragment.java
@@ -1,0 +1,36 @@
+package com.example.KC.ui.bottom.search;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.example.KC.R;
+
+public class SearchFragment extends Fragment {
+
+    private SearchViewModel mViewModel;
+
+    public static SearchFragment newInstance() {
+        return new SearchFragment();
+    }
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.search_fragment, container, false);
+    }
+
+    @Override
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        mViewModel = new ViewModelProvider(this).get(SearchViewModel.class);
+        // TODO: Use the ViewModel
+    }
+
+}
