+++ app/src/main/java/com/example/KC/ui/bottom/search/SearchViewModel.java
@@ -1,0 +1,7 @@
+package com.example.KC.ui.bottom.search;
+
+import androidx.lifecycle.ViewModel;
+
+public class SearchViewModel extends ViewModel {
+    // TODO: Implement the ViewModel
+}
