+++ app/src/main/res/layout/fragment_money.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.money.MoneyFragment">
+    tools:context=".ui.side.money.MoneyFragment">
 
     <TextView
         android:id="@+id/link_money"
