+++ app/src/main/res/layout/fragment_news.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.news.NewsFragment">
+    tools:context=".ui.bottom.news.NewsFragment">
 
     <TextView
         android:id="@+id/news"
