+++ app/src/main/res/layout/fragment_time.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.time.TimeFragment">
+    tools:context=".ui.side.time.TimeFragment">
 
     <TextView
         android:id="@+id/link_time"
